+++ tests_smoke/smoke_test.py
@@ -1,11 +1,11 @@
+from test_admin_email_csv import test_admin_email_csv
 from test_admin_email_one_off import test_admin_email_one_off
+from test_admin_sms_csv import test_admin_sms_csv
 from test_admin_sms_one_off import test_admin_sms_one_off
 from test_api_email import test_api_email
-from test_api_sms import test_api_sms
 from test_api_email_bulk import test_api_email_bulk
+from test_api_sms import test_api_sms
 from test_api_sms_bulk import test_api_sms_bulk
-from test_admin_email_csv import test_admin_email_csv
-from test_admin_sms_csv import test_admin_sms_csv
 
 if __name__ == "__main__":
     test_api_email()
