+++ tests_smoke/test_admin_email_csv.py
@@ -1,9 +1,16 @@
-from common import job_line, rows_to_csv, s3upload, set_metadata_on_csv_upload, get_csv_location, pretty_print
+import os
+import time
+
+import requests
+from common import (
+    job_line,
+    pretty_print,
+    rows_to_csv,
+    s3upload,
+    set_metadata_on_csv_upload,
+)
 from dotenv import load_dotenv
 from notifications_python_client.authentication import create_jwt_token
-import os
-import requests
-import time
 
 load_dotenv()
