+++ tests_smoke/test_api_sms.py
@@ -1,30 +1,30 @@
+import os
+import time
+
+import requests
 from common import pretty_print
 from dotenv import load_dotenv
-import json
-import os
-import requests
-import time
 
 load_dotenv()
 
 API_KEY = os.environ.get("API_KEY")
-EMAIL_TEMPLATE_ID = os.environ.get("EMAIL_TEMPLATE_ID")
-EMAIL_TO = os.environ.get("EMAIL_TO")
+SMS_TEMPLATE_ID = os.environ.get("SMS_TEMPLATE_ID")
+SMS_TO = os.environ.get("SMS_TO")
 API_HOST_NAME = os.environ.get("API_HOST_NAME")
 
 
-def test_api_email():
-    print("test_api_email... ", end="", flush=True)
+def test_api_sms():
+    print("test_api_sms... ", end="", flush=True)
     response = requests.post(
-        f"{API_HOST_NAME}/v2/notifications/email",
+        f"{API_HOST_NAME}/v2/notifications/sms",
         json={
-            "email_address": EMAIL_TO,
-            "template_id": EMAIL_TEMPLATE_ID,
+            "phone_number": SMS_TO,
+            "template_id": SMS_TEMPLATE_ID,
         },
         headers={"Authorization": f"ApiKey-v1 {API_KEY[-36:]}"},
     )
     if response.status_code != 201:
-        print("FAILED: post to v2/notifications/email failed")
+        print("FAILED: post to v2/notifications/sms failed")
         pretty_print(response.json())
         return
 
@@ -40,7 +40,7 @@ def test_api_email():
             break
 
     if response.status_code != 200:
-        print("FAILED: email not sent successfully")
+        print("FAILED: sms not sent successfully")
         pretty_print(response.json())
         return
 
@@ -48,4 +48,4 @@ def test_api_email():
 
 
 if __name__ == "__main__":
-    test_api_email()
+    test_api_sms()
